+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix tall flowers not being able to be used to craft dyes
 - Add Forager Trades and Config
 - Add Vanilla Trades and Disable Option
+- Fix WhitePuffBall age not being reset when picked, Don't call super.use on ages less than 1
 
 # 1.1.2-Beta
 - Set getBucketItemStack in ManOWar.java to correct Item