+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add Forager Trades and Config
 - Add Vanilla Trades and Disable Option
 - Fix WhitePuffBall age not being reset when picked, Don't call super.use on ages less than 1
+- Fix WhitePuffBall growth being incorrect
 
 # 1.1.2-Beta
 - Set getBucketItemStack in ManOWar.java to correct Item